Added More Cars to the Traffic
Added more cars to the traffic and added the ability to be controlled by the neural network just by giving the parameter "AI" to the car object while creation.
Similarly, passing "DUMMY" will make the act as a dummy by just going at a set speed in the forward direction.
And passing "KEYS" will allow the user to control the car using the arrow keys.